// frontend/src/api/model/composeServiceOrderRequest.ts
/**
 * Generated by orval v8.17.0 🍺
 * Do not edit manually.
 * KS Tech Service Management API
 * API for managing users, Products, and service orders in a service-center / repair business. Designed using API-first principles.
 * OpenAPI spec version: 1.0.0
 */
import type { ComposeContactInput } from './composeContactInput';
import type { ComposeCustomerInput } from './composeCustomerInput';
import type { ComposeProductInput } from './composeProductInput';
import type { ComposeServiceOrderInput } from './composeServiceOrderInput';
import type { ComposeUserProductInput } from './composeUserProductInput';
import type { ExistingEntityReferences } from './existingEntityReferences';

export interface ComposeServiceOrderRequest {
  existing: ExistingEntityReferences;
  customer: ComposeCustomerInput;
  contact: ComposeContactInput;
  product: ComposeProductInput;
  userProduct: ComposeUserProductInput;
  serviceOrder: ComposeServiceOrderInput;
}